Add some parameters to enable managesieve
This patch add 3 parameters in order to control activation of
managesieve plugin on dovecot.
Default behavior is to not enable managesieve protocol so that dovecot
does not listen unexpectidely on port 4190.
If enabled, managesieve will listen for ManageSieve commands.
Also install dovecot-managesieved package on Debian like OS as it was
missing.